requirements/prod.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/prod.txt requirements/prod.in
#
asgiref==3.11.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
boto3==1.42.57
    # via -r requirements/base.in
botocore==1.42.57
    # via
    #   boto3
    #   s3transfer
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.5
    # via pyjwt
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==5.2.11
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   django-ckeditor
    #   django-cors-headers
    #   django-filter
    #   django-js-asset
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   social-auth-app-django
django-allauth==65.14.3
    # via -r requirements/base.in
django-ckeditor==6.7.3
    # via -r requirements/base.in
django-cors-headers==4.9.0
    # via -r requirements/base.in
django-environ==0.13.0
    # via -r requirements/base.in
django-filter==25.2
    # via -r requirements/base.in
django-js-asset==3.1.2
    # via django-ckeditor
django-storages==1.14.6
    # via -r requirements/base.in
djangorestframework==3.16.1
    # via
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via -r requirements/base.in
drf-yasg==1.21.15
    # via -r requirements/base.in
gunicorn==25.1.0
    # via -r requirements/prod.in
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-yasg
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
oauthlib==3.3.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==26.0
    # via
    #   drf-yasg
    #   gunicorn
pillow==12.1.1
    # via -r requirements/base.in
psycopg2-binary==2.9.11
    # via -r requirements/base.in
pycparser==3.0
    # via cffi
pyjwt[crypto]==2.11.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-dateutil==2.9.0.post0
    # via botocore
python3-openid==3.2.0
    # via social-auth-core
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.3
    # via drf-yasg
requests==2.32.5
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
social-auth-app-django==5.7.0
    # via -r requirements/base.in
social-auth-core==4.8.5
    # via social-auth-app-django
sqlparse==0.5.5
    # via django
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.6.3
    # via
    #   botocore
    #   requests
whitenoise==6.11.0
    # via -r requirements/prod.in
